pyproject.toml
[project]
name = "cron-job"
version = "0.1.0"
description = ""
authors = [
    {name = "mahmud",email = "mahmud@nybsys.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "python-crontab (>=3.3.0,<4.0.0)",
    "croniter (>=6.0.0,<7.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
